requirements.txt
boto3==1.39.1
langgraph==0.6.7
openai==1.93.0
requests==2.32.4
mcp==1.12.4
psycopg2-binary==2.9.10
pgvector==0.4.1
Jinja2==3.1.6
python-json-logger==3.3.0
uvicorn==0.34.2
httpx==0.28.1
a2a-sdk[postgresql]==0.3.7
asyncpg==0.30.0
ruff>0.4.0
grandalf==0.8
jsonpath-ng==1.7.0
jsonpath2==0.4.5